+++ R/grid_area.R
@@ -24,14 +24,13 @@
 #'   plotting purposes.
 #'   
 #' @examples
-#' library(sf)
-#'
 #' ## Exemplary input Polygon with 2km x 2km:
+#' library(sf)
 #' Polygon1 <- sf::st_as_sf(sf::st_sfc(
-#'   sf::st_polygon(list(cbind(
-#'     c(4498482, 4498482, 4499991, 4499991, 4498482),
-#'     c(2668272, 2669343, 2669343, 2668272, 2668272)))), 
-#'   crs = 3035
+#'  sf::st_polygon(list(cbind(
+#'    c(0, 0, 2000, 2000, 0),
+#'    c(0, 2000, 2000, 0, 0)))),
+#' crs = 3035
 #' ))
 #'
 #' ## Create a Grid