+++ R/windfarmGA_package.R
@@ -7,9 +7,9 @@
 #' A package to optimize small wind farms with irregular shapes
 #' using a genetic algorithm. It requires a fixed amount of turbines, a fixed
 #' rotor radius and an average wind speed value for each incoming wind
-#' direction. A terrain effect model can be included that downloads an 'SRTM'
-#' elevation model and loads a Corine Land Cover raster to approximate surface
-#' roughness. Further information can be found in the description of the
+#' direction. A terrain effect model can be included which downloads a digital
+#' elevation model and a Corine Land Cover raster to approximate surface
+#' roughness. Further information can be found at the description of the
 #' function \code{\link{windfarmGA}}.
 #' 
 #' @seealso 
@@ -34,21 +34,13 @@
 #'   layout axis legend mtext
 #' @importFrom RColorBrewer brewer.pal
 #' @importFrom calibrate textxy
-#' @importFrom ggplot2 ggplot geom_tile geom_point geom_bar aes
-#'   scale_fill_gradient coord_equal labs  scale_x_discrete waiver coord_polar
-#'   scale_fill_manual theme element_blank ylim
-#' @importFrom leaflet colorFactor iconList makeIcon leaflet addTiles
-#'   addProviderTiles popupOptions addMarkers addCircleMarkers hideGroup
-#'   addPolygons addLegend labelFormat addLayersControl layersControlOptions
-#'   addPopups
 #' @importFrom stats runif smooth.spline sd aggregate median dist complete.cases
 #'   quantile
 #' @importFrom utils download.file unzip read.csv globalVariables
 #' @importFrom doParallel registerDoParallel
 #' @importFrom parallel makeCluster stopCluster detectCores
 #' @importFrom foreach foreach %dopar% 
-#' @importFrom methods as slot slot<-
-#' @importFrom gstat idw
+#' @importFrom methods as
 #' @importFrom magrittr %>% 
 #' @importFrom Rcpp sourceCpp
 #'