+++ data-raw/create_data.R
@@ -1,7 +1,8 @@
 ## Create some Polygons ###################
-Projection <- '+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000
-+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs'
-Proj84 <- '+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'
+library(sp)
+library(sf)
+Projection <- 3035
+Proj84 <- 4326
 
 cordslist <- list(structure(c(15.8862501316651, 15.8802827241768, 15.8814016130809, 
                  15.9090008727144, 15.9168330950428, 15.9265301322114, 15.9608427252693, 
@@ -13,10 +14,14 @@ cordslist <- list(structure(c(15.8862501316651, 15.8802827241768, 15.88140161308
                  47.4918468635213, 47.4908388168251, 47.4739511582908, 47.4794969582227, 
                  47.4986506726316, 47.5006664467725, 47.5273681479331, 47.5505321961794, 
                  47.5568250061956), .Dim = c(18L, 2L)))
-big_shape = SpatialPolygons(list(Polygons(list(Polygon(cordslist)), 1)), 
-                            proj4string = CRS(Proj84))
-big_shape <- spTransform(big_shape, CRS(Projection))
+big_shape <- sf::st_as_sf(sf::st_sfc(
+  sf::st_polygon(cordslist), 
+  crs = Proj84
+))
+big_shape <- st_transform(big_shape, st_crs(Projection))
 usethis::use_data(big_shape, overwrite = TRUE)
+
+
 
 cordslist <- list(structure(c(15.0817690096444, 15.0830743800325, 15.0950091950091, 
                               15.0976199357853, 15.0862445652606, 15.0771069725441, 15.0817690096444, 
@@ -33,15 +38,38 @@ cordslist <- list(structure(c(15.0817690096444, 15.0830743800325, 15.09500919500
                               15.1308136399391, 47.1309979565454, 47.1202130906391, 47.1161523392859, 
                               47.1148832909329, 47.1176751573605, 47.1250348303582, 47.1318860245243, 
                               47.1345501394874, 47.1344232798964, 47.1309979565454), .Dim = c(10L, 2L)))
-cordslist <- lapply(cordslist, Polygon) 
-multi_shape = SpatialPolygons(list(Polygons(cordslist, 1)),
-                             proj4string = CRS(Proj84))
-multi_shape <- spTransform(multi_shape, CRS(Projection))
+multi_shape <- sf::st_as_sf(sf::st_sfc(
+  sf::st_polygon(cordslist), 
+  crs = Proj84
+))
+multi_shape <- st_transform(multi_shape, st_crs(Projection))
 usethis::use_data(multi_shape, overwrite = TRUE)
 
-library(rgdal)
-hole_shape <- readOGR("./inst/extdata/hole_shape.shp")
-hole_shape <- spTransform(hole_shape, CRS(Projection))
+
+
+cordlist <- structure(c(4585270.60241151, 4584979.13247779, 4585868.18527401, 
+                        4586290.18668221, 4587632.0127656, 4588005.31974706, 4587933.05332353, 
+                        4586395.52121144, 4586361.43925126, 4585360.80689825, 4585175.08261164, 
+                        4584678.33214126, 4584402.04363248, 4584174.83760134, 4585270.60241151, 
+                        4585976.13994593, 4585791.24297194, 4586035.56149181, 4586465.59095286, 
+                        4586802.20350957, 4586755.56930998, 4585976.13994593, 2655882.39876357, 
+                        2656642.18821161, 2655893.72473733, 2655340.35909448, 2654724.59708819, 
+                        2654018.63277193, 2652784.8782917, 2652651.15264346, 2653278.66439827, 
+                        2653888.59490599, 2654286.23636729, 2654391.02616789, 2655062.69468666, 
+                        2655886.82997589, 2655882.39876357, 2654869.24044678, 2654754.8149049, 
+                        2654298.7450985, 2653940.44349285, 2654284.70017805, 2654449.48867358, 
+                        2654869.24044678, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
+                        2, 2, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
+                        1, 1, 1, 1, 1, 1, 1, 1), .Dim = c(22L, 4L), .Dimnames = list(
+                          NULL, c("X", "Y", "L1", "L2")))
+df <- data.frame(cordlist)
+df <- split(df[,1:2], df[,"L1"])
+names(df) <- NULL
+hole_shape <- sf::st_as_sf(sf::st_sfc(
+  sf::st_polygon(lapply(df, as.matrix)), 
+  crs = Proj84
+))
+# plot(hole_shape, col="red")
 usethis::use_data(hole_shape, overwrite = TRUE)
 
 ## Download Corine-Land-Cover Raster ######################
@@ -71,13 +99,12 @@ usethis::use_data(k_weibull, overwrite = TRUE)
 
 
 ## Result with Rect and 200 Iteration ##################
-sp_polygon <- Polygon(rbind(c(4498482, 2668272), c(4498482, 2669343),
-                            c(4499991, 2669343), c(4499991, 2668272)))
-sp_polygon <- Polygons(list(sp_polygon), 1)
-sp_polygon <- SpatialPolygons(list(sp_polygon))
-projection <- paste("+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000",
-                    "+ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")
-proj4string(sp_polygon) <- CRS(projection)
+sp_polygon <- sf::st_as_sf(sf::st_sfc(
+  sf::st_polygon(list(cbind(
+    c(4498482, 4498482, 4499991, 4499991, 4498482),
+    c(2668272, 2669343, 2669343, 2668272, 2668272)))),
+  crs = 3035
+))
 usethis::use_data(sp_polygon, overwrite = TRUE)
 
 winddat <- data.frame(ws = 12, wd = 0)