+++ tests/testthat/test_ga.R
@@ -20,7 +20,7 @@ test_that("Test Genetic Algorithm with different Inputs", {
   ))
   data.in <- data.frame(ws = 12, wd = 0)
 
-  ## SpatialPolygon Input - 100 Iteration #####################
+  ## SF Polygon Input - 100 Iteration #####################
   resultSP <- expect_warning(quiet(genAlgo(Polygon1 = Polygon1,
                       n = 20, iteration = 100,
                       vdirspe = data.in,
@@ -105,8 +105,7 @@ test_that("Test Genetic Algorithm with different Inputs", {
   expect_false(any(unlist(sapply(resultDF, is.na))));
 
   ## Matrix Input #####################
-  PolygonMat <- st_coordinates(Polygon1)
-  PolygonMat <- as.matrix(PolygonMat[,1:2])
+  PolygonMat <- as.matrix(PolygonDF[,1:2])
   resultMA <- expect_warning(genAlgo(Polygon1 = PolygonMat, plotit = TRUE,
                         n = 20, iteration = 1,
                         vdirspe = data.in,
@@ -128,8 +127,6 @@ test_that("Test Genetic Algorithm with different Inputs", {
   # expect_true(any(as.vector(sapply(resultMA100[,3], function(x) x[,"EfficAllDir"] == 100))))
 
   ## Test with non default arguments ####################
-  PolygonMat <- st_coordinates(Polygon1)
-  PolygonMat <- as.matrix(PolygonMat[,1:2])
   colnames(PolygonMat) <- c("hor", "vert")
   resultMA <- expect_warning(genAlgo(Polygon1 = PolygonMat,
                       n = 20, iteration = 1, GridMethod = "h",
@@ -143,16 +140,13 @@ test_that("Test Genetic Algorithm with different Inputs", {
   expect_false(any(unlist(sapply(resultMA, is.na))))
   
   resultMA <- expect_warning(genAlgo(Polygon1 = PolygonMat,
-                      n = 20, iteration = 1, GridMethod = "h",
+                      n = 15, iteration = 1, GridMethod = "h",
                       vdirspe = data.in, elitism = T, nelit = 10000, 
                       selstate = "var", crossPart1 = "ran", 
                       trimForce = TRUE, mutr = 15,
-                      Rotor = 50, 
-                      # Projection = "+proj=tmerc +lat_0=0 +lon_0=31 +k=1 +x_0=0 +y_0=-5000000 +ellps=bessel +pm=ferro +units=m +no_defs",
-                      Projection = "+proj=lcc +lat_1=49 +lat_2=46 +lat_0=47.5 +lon_0=13.33333333333333 +x_0=400000 +y_0=400000 +ellps=bessel +towgs84=577.326,90.129,463.919,5.137,1.474,5.297,2.4232 +units=m +no_defs",
-                      # Projection = "+proj=tmerc +lat_0=0 +lon_0=28 +k=1 +x_0=0 +y_0=0 +ellps=bessel +pm=ferro +units=m +no_defs",
-                      # Projection = "+proj=lcc +lat_1=49 +lat_2=46 +lat_0=47.5 +lon_0=13.33333333333333 +x_0=400000 +y_0=400000 +ellps=GRS80 +units=m +no_defs",
-                      # Projection = "+proj=tmerc +lat_0=0 +lon_0=15 +k=1 +x_0=6500000 +y_0=0 +ellps=bessel +units=m +no_defs",
+                      Rotor = 30, 
+                      # Projection = "+proj=lcc +lat_1=49 +lat_2=46 +lat_0=47.5 +lon_0=13.33333333333333 +x_0=400000 +y_0=400000 +ellps=bessel +towgs84=577.326,90.129,463.919,5.137,1.474,5.297,2.4232 +units=m +no_defs",
+                      Projection = 3035,
                       RotorHeight = 100, plotit = F))
   expect_true(nrow(resultMA) == 1)
   expect_is(resultMA, "matrix")